backend/weather_prediction/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
numpy
pydantic
pydantic-settings
aiohttp
celery
httpx
apscheduler
kafka-python
redis[hiredis]
scikit-learn